requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
faiss-cpu
sentence-transformers
python-dotenv
pypdf
python-multipart
fastapi
uvicorn[standard]
pydantic
